+++ backend/canisters/group/CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
-- When a member is removed also remove any invite ([#8328](https://github.com/open-chat-labs/open-chat/pull/8328))
+- Call to remove member will also remove any invite ([#8332](https://github.com/open-chat-labs/open-chat/pull/8332))
 
 ## [[2.0.1828](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1828-group)] - 2025-07-09